+++ tac/tac.c
@@ -203,7 +203,7 @@ struct TAC* makeTACUnaryOp(uint32_t dest, uint32_t src, enum TAC_OP op) {
 	return t;
 }
 
-struct TAC* makeTACStoreConstAddr(uint32_t addr, uint32_t src) {
+struct TAC* makeTACStoreConstAddr(uint64_t addr, uint64_t src) {
 
 	struct TAC* t = makeTAC();
 	t->kind = TAC_STORE_CONST_ADDR;
